Try using the word highlight shortcuts on the arrow keys
In theory, this will let me use the arrow keys on the NAV layer to
highlight words for copying and pasting.

Will this work better in practice than the previous location which was
Vim keys offset by one row higher?  Time will tell.

I'm slightly optimistic because I've gotten used to the arrow keys and
this might offer a slight optimization over Shift+Arrow keys. But we'll
see.  I've become pretty fast with Shift+Arrow so far. And the NAV layer
key is not in a convenient place.

Open question: would I use the NAV layer more if the layer key was more
convenient? Like somwhere much closer to the thumb?