Remove dual function for Left ALT key
It turns out that having ALT/Option on the left side has been too useful
for two-handed combos not to prioritize the key being dedicated and not
shared with another.

It was getting very annoying to share it with a key I used as little as
an asterisk, so let's just make it a full-time ALT now.